fix problematic fixtures
