+++ tests/test_user_and_admin_user_fixture.py
@@ -1,0 +1,2 @@
+def test_user_and_admin_user_fixture_coexist_nicely(user, admin_user):
+    pass
